+++ test/function/samples/default-exports-in-parens-class-with-id/_config.js
@@ -1,0 +1,3 @@
+module.exports = {
+	description: 'handles default class with id exports wrapped in parens'
+};