+++ test/function/samples/default-exports-in-parens-class-with-id/main.js
@@ -1,0 +1,3 @@
+import Foo from './foo.js';
+var obj = new Foo();
+assert.ok( obj instanceof Foo );