+++ test/function/samples/default-exports-in-parens-function-without-id/_config.js
@@ -1,0 +1,3 @@
+module.exports = {
+	description: 'handles default function without id exports wrapped in parens'
+};